Add motion profiling for snap-to-angle reqests to teleop robot orientation driver (#43)
* Add motion profiling for snap-to-angle reqests to teleop robot orientation driver

* Cleanup and comments

* Add config items and ddr for profile max speed, accel, jerk
Minor fixes for normalizing angles

* Debugging

* Fixed transition from joystick rotation to profile turning by forcing initial profile velocity to 0

* Fixed config file param name

* Stop building teleop for the rio, no ruckig lib there (eventually turn back on after fixing this in the ROS rio build)

* reenable teleop build on rio after docker update

* Static analysis fixes

* Fixes found by static analysis

* Cleanup

+++ zebROS_ws/src/frc_interfaces/include/frc_interfaces/robot_controller_interface.h
@@ -131,7 +131,7 @@ class RobotControllerState
 		double      current6v_{0};
 		bool        enabled6v_{0};
 		int         fault_count6v_{0};
-		float       can_percent_bus_utilization_{0};
+		double      can_percent_bus_utilization_{0};
 		int         can_bus_off_count_{0};
 		int         can_tx_full_count_{0};
 		int         can_receive_error_count_{0};
